# specs/openapi.yaml
openapi: 3.0.3
info:
  title: FaMoney API
  description: 공동 지출 원장 웹 애플리케이션 API
  version: 1.0.0
  contact:
    name: FaMoney Team
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.famoney.app
    description: Production server
...
paths:
...
  /api/v1/ledgers:
    post:
      tags: [Ledgers]
      summary: 원장 생성
      operationId: createLedger
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLedgerRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Ledgers]
      summary: 내 원장 목록 조회
      operationId: getMyLedgers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/ledgers/{ledgerId}:
    get:
      tags: [Ledgers]
      summary: 원장 상세 조회
      operationId: getLedger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Ledgers]
      summary: 원장 수정
      operationId: updateLedger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLedgerRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Ledgers]
      summary: 원장 삭제
      operationId: deleteLedger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Members
  # ============================================
  /api/v1/ledgers/{ledgerId}/members:
    get:
      tags: [Members]
      summary: 멤버 목록 조회
      operationId: getLedgerMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ledgers/{ledgerId}/members/invite:
    post:
      tags: [Members]
      summary: 멤버 초대
      operationId: inviteMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '201':
          description: 초대 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ledgers/{ledgerId}/members/{memberId}:
    patch:
      tags: [Members]
      summary: 멤버 역할 변경
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/MemberIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Members]
      summary: 멤버 제거
      operationId: removeMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/MemberIdParam'
      responses:
        '204':
          description: 제거 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================
  # Parameters
  # ============================================
  parameters:
    LedgerIdParam:
      name: ledgerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 원장 ID
    MemberIdParam:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 멤버 ID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: 페이지 번호 (0부터 시작)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: 페이지 크기
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    # ------------------------------------------
    # Auth
    # ------------------------------------------
...
    CreateLedgerRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        currency:
          type: string
          default: KRW
          enum: [KRW, USD, JPY, EUR]
    UpdateLedgerRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
    LedgerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        currency:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: '#/components/schemas/UserSummary'
    LedgerDetailResponse:
      allOf:
        - $ref: '#/components/schemas/LedgerResponse'
        - type: object
          properties:
            myRole:
              $ref: '#/components/schemas/MemberRole'
            totalExpenseAmount:
              type: number
              format: double
            recentExpenses:
              type: array
              items:
                $ref: '#/components/schemas/ExpenseSummary'
    LedgerListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/LedgerResponse'
        page:
          $ref: '#/components/schemas/PageInfo'

    # ------------------------------------------
    # Member
    # ------------------------------------------
    MemberRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER, VIEWER]
    InviteMemberRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/MemberRole'
    UpdateMemberRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          $ref: '#/components/schemas/MemberRole'
    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserSummary'
        role:
          $ref: '#/components/schemas/MemberRole'
        joinedAt:
          type: string
          format: date-time
    MemberListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MemberResponse'
    InvitationResponse:
      type: object
      properties:
        inviteCode:
          type: string
        inviteUrl:
          type: string
        expiresAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        profileImage:
          type: string
          nullable: true

    # ------------------------------------------
    # Expense
    # ------------------------------------------
...
    ExpenseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        description:
          type: string
        category:
          type: string
          nullable: true
        expenseDate:
          type: string
          format: date
...